requirements-api.txt
flask==2.3.3
flask-cors==4.0.0
python-dotenv==1.0.0

# For the main agent system (already in requirements.txt)
# google-adk-agents
# tachyon-adk-client
# etc.
